doubly_linked_list.cpp Add Doubly Linked List
This commit implements a doubly linked list which is one of the most popular data structures. A doubly linked list is similar to a linked list with an additional pointer which points to the previous node. Thus, elements can be traversed in both the directions. This commit  consists of a list of common operations performed on a doubly linked list.

Closes https://github.com/NITSkmOS/Algorithms/issues/32