extended_euclidean_gcd.py: Add Extended Euclidean GCD Algorithm
This adds Extended Euclidean Algorithm with Python implementation of the same.
This implementation uses iterative method of the algorithm, takes two
values and gives out the gcd along with the polynomial equation to get
inverse directly if the gcd is 1. Code uses generic notations and
naming standard used in documentation along with that found in Wikipedia.

Closes https://github.com/NITSkmOS/Algorithms/issues/3